% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{nrouter_embeddings}
\alias{nrouter_embeddings}
\title{POST /embeddings}
\usage{
nrouter_embeddings(client, body)
}
\arguments{
\item{client}{An \code{nrouter_client}.}

\item{body}{Named list forming the request body.}
}
\value{
A list with \code{body}, \code{meta} and \code{status_code}.
}
\description{
POST /embeddings
}
