% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{nrouter_model}
\alias{nrouter_model}
\title{GET /models/(model_id)}
\usage{
nrouter_model(client, model_id)
}
\arguments{
\item{client}{An \code{nrouter_client}.}

\item{model_id}{Model identifier. It is encoded as one URL path segment.}
}
\value{
A list with \code{body}, \code{meta} and \code{status_code}.
}
\description{
GET /models/(model_id)
}
